Pase las carpetas transporte y solicitudes a la raiz del repo

+++ Solicitudes/src/main/java/com/microservicios/Solicitudes/dto/request/CrearSolicitudDTO.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CrearSolicitudDTO {
     
-    private Integer idCliente;
-    private Integer idContenedor;
+    private Long idCliente;
+    private Long idContenedor;
 }
