+++ Solicitudes/src/main/java/com/microservicios/Solicitudes/dto/responses/SolicitudDTO.java
@@ -3,7 +3,6 @@ package com.microservicios.Solicitudes.dto.responses;
 import java.time.LocalDate;
 import com.microservicios.Solicitudes.entity.EstadoSolicitud;
 import com.microservicios.Solicitudes.entity.Ruta;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,8 +13,8 @@ import lombok.NoArgsConstructor;
 public class SolicitudDTO {
 
     private Integer id;
-    private Integer idContenedor;
-    private Integer idCliente;
+    private Long idContenedor;
+    private Long idCliente;
     private Double costoEstimado;
     private Double costoReal;
     private String tiempoEstimado;
